renaming shared lib to libmemex
