+++ lib/libmemex/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "shared"
+name = "libmemex"
 version = "0.1.0"
 edition = "2021"
