+++ lib/libmemex/src/lib.rs
