+++ lib/libmemex/src/vector/local.rs
