+++ lib/libmemex/src/vector/mod.rs
