+++ lib/worker/Cargo.toml
@@ -13,5 +13,5 @@ sea-orm = { workspace = true }
 serde_json = { workspace = true }
 uuid = { version = "1.3.1", default-features = false, features = ["serde", "v5"] }
 embedder = { path = "../embedder" }
-shared = { path = "../shared" }
+libmemex = { path = "../libmemex" }
 url = "2.4.0"
